Stop laughing
I managed to remember to NOT put integers as strings before uploading this commit :P  Anyway, I've managed to get the code to pull the high bidder!  Now I need to figure out how to display the value.  I THINK it's as simple as calling value = instead of key = but we'll see! <3