Reverted main.py
So I was thinking about it, and I realized using a built in function like max() isn't really in the spirit of what I've been learning about so far.  I think the intent was to use a for loop instead, or "I have to do this the hard way" lol.  Ill do some more reading and see if I can't figure it out using only existing code foundations :)